backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Cors==3.0.10
python-dotenv==1.0.0
openai==1.9.0
